server/requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
pytesseract
Pillow
pydantic-settings
python-dotenv
pymupdf
python-docx
pandas
openpyxl
langchain-text-splitters
google-generativeai
scikit-learn
numpy
